refactor(backend): index.ts 54 handler を handlers/ 配下に機能領域別に分離 (#1144 Phase-1)
#1144 提案 A の実装。index.ts に同居していた 64 個の MCP tool ハンドラ
(designer__* 54 件 + editSession__* 10 件) を、機能領域別に
handlers/<feature>.ts に分割した。

新規 handler:
- handlers/designer.ts    — GrapesJS キャンバス操作 (8)
- handlers/screen.ts      — Screen CRUD + export (5)
- handlers/edge.ts        — 画面フロー (4)
- handlers/customBlock.ts — カスタムブロック (3)
- handlers/puckComponent.ts — Puck カスタムコンポーネント (3)
- handlers/table.ts       — テーブル定義 CRUD (5)
- handlers/pageLayout.ts  — PageLayout CRUD + get/save (6)
- handlers/export.ts      — DDL/spec/ER 生成 (4)
- handlers/tab.ts         — タブ管理 + 保存 (6)
- handlers/screenItem.ts  — 画面項目 rename (4)
- handlers/workspace.ts   — workspace 管理 (6)
- handlers/editSession.ts — editSession (10)
- handlers/ddl.ts         — DDL 生成ヘルパー (純粋関数)

index.ts は以下に縮退:
- MCP server entry + lifecycle
- HTTP transport mount
- ハンドラ配列 dispatcher (loop で順次試行)
- 共通エラーハンドリング

各 handler は既存の ToolHandler signature (name/args/root/sessionId) を
踏襲し、該当しない tool は null を返して dispatcher に次を試させる。
既存の handlers/processFlow.ts / handlers/marker.ts と同じ pattern。

LOC: index.ts 1830 → 235 (-1595, -87%)、新規 handlers 計 ~2,015 LOC
(comment header + 1 file 1 領域の構造化分のオーバーヘッド)。

不変:
- MCP tool 名 (designer__* / editSession__*) は変更なし — AI client 契約
- RPC method 名 (wsBridge cmd) は変更なし — frontend 契約
- schema 無変更 (schemas/)
- 既存テスト 513 件すべて pass、frontend build pass、backend build pass

#1144 提案 (B) wsBridge 分割 / (C) RPC 命名整理 は Phase-2/-3 として
ISSUE 本文の sub-section で trace 継続 (新規 ISSUE 起票はしない)。

Refs #1144
